Unit tests for 'provider/http'.
